+++ frontend/app/src/main/java/com/example/frontend/model/AuthResponse.kt
@@ -6,5 +6,6 @@ data class AuthResponse(
     val token: String,
     val userName: String,
     val userCurrentLocation: LatLng?,
-    val userMail: String
+    val userMail: String,
+    val userProfile: Int
 )
